Fix true_divide unit test on macOS with C (#2388)
When a very large complex number is divided by itself, the result should
be `1` as long as the number is representable in finite precision. In
practice, due to the way complex division is implemented, an overflow in
the real part happens for numbers larger (in absolute value) than a
certain value.

We consider complex numbers of the form `c = f * (1 + 1j)`, where `f` is
a 64-bit floating-point number. Our experiments have found that `c / c
== 1` as long as `min_float/2 <= f <= max_float/2`, where `min_float =
numpy.finfo('float').min` and `max_float = numpy.finfo('float').max` are
the minimum and maximum representable floating-point numbers,
respectively.

An exception to the rule above has been found on macOS when the C
language was selected. In this case the bounds must be further reduced,
dividing them by `sqrt(2)`. We apply this correction to the unit test
function `test_true_divide` in
`tests/epyccel/recognised_functions/test_numpy_funcs.py`. This fixes
#2391.