Rename the variable `accelerators` in the compilation code (#2399)
In the code the word `accelerators` is used to hold all keys which
describe tools that are used and require extra compilation flags/include
dirs/libs/etc. The majority of these tools are indeed accelerators
(OpenMP, OpenACC, MPI), but Python is also one such tool as extra libs
and flags are required to build the `.so` Python library. This PR
renames `accelerators` -> `extra_compilation_tools` where appropriate in the code to
reflect this role more accurately.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>