Add a `Wrappergen` class to reduce code duplication (#2386)
Add a `Wrappergen` class. This homogenises the treatment of the wrapper
stage. It is important to reduce code duplication when adding new
printers (e.g. a C++ printer which will not need to use C for the
interface) or when adding new compilation methods (see #221).

**Commit Summary**
- Remove unused `target_language` parameter from `CWrapperCodePrinter`
so all printers have the same constructor arguments
- Rename badly named variable `file_location` -> `sharedlib_dirpath` in
wrapper classes
- Define `start_language` and `target_language` class variables in
wrapper subclasses
- Create a `Wrappergen` class to group calls to `Wrapper.wrap` methods
and `Printer.print` methods during wrapper generation
- Use `Wrappergen` in `create_shared_library` method