Fix bug when inlining a call to a function with no return, via a dotted function call (#2409)
The method `_visit_InlineFunctionCall` assumes that a `FunctionCall` is
either in an `Assign` node due to the anonymous assign nodes created
here:

https://github.com/pyccel/pyccel/blob/55adae9a4c0ae39f75392b16cc5be34d815bef83/pyccel/parser/syntactic.py#L489-L495

or is in a statement (e.g. `f() + 3`). In the second case the function
must return something. This means that `_visit_InlineFunctionCall`
assumes that the function always returns something if there is no assign
node.

This assumption is violated in the case of a dotted function call as the
code creates a temporary node for the function call here:

https://github.com/pyccel/pyccel/blob/6e00555d2b3d3a5fc06ec248579bb88f027de018/pyccel/parser/semantic.py#L3358
This node has no parents and is therefore not found in an `Assign` node.
The easy fix for this problem is to add the missing parent. Fixes #2408

**Commit Summary**
- Add "inlining" to dictionary
- Ensure `FunctionCall` has user nodes
- Remove dead code in syntactic stage which leads to annotations having
multiple user nodes
- Add missing `FunctionDefResult` variable to `InlineFunctionDef`
created to represent a lambda function.
- Add tests for bug