Fix wrapper bug when returning an instance of a class (#2423)
Fix wrapper bug when returning an instance of a class from the method of
a preceding class. Fixes #2248

The issue arose due to the fact that the type had not yet been declared
at the moment it was used.

Additionally the type of an inhomogeneous tuple is fixed. The use of the
wrong type caused a warning but did not cause bugs as the pointer
pointed to the correct type.