Document and uniformise scope keys (#2413)
According to the developer docs, the keys for the dictionaries in the
scope are the names of the original Python objects. However this does
not always seem to be the case. Particularly in the case of functions
and interfaces where the functions contained in the interfaces were
inserted into the scope under their low-level name. As a result it is
easy to make mistakes when handling objects that have been renamed (as
seen in #2410).

This PR aims to improve this situation by giving `Scope` complete
control over its contents. Objects are always inserted using an
`insert_X` function which enforces that original Python names are used
as keys. A class `ReadOnlyDict` is introduced to prevent bypassing these
`insert_X` methods.

Storing functions in `scope.functions` under their original Python name
means that interface internal functions are no longer stored in this
dictionary. As a result they are also not stored directly in a `Module`
object. The logic of this is more intuitive but it has knock-on effects
in the printing stage where the functions must be extracted from the
interface.

**Commit Summary**
- Add new dependency on `immutabledict`
- Add more detail to developer docs
- Update `CCodePrinter`
    - Print interface functions via `_print_Interface`
- Extract interface functions from the Interface for the header printing
- Remove excessive vertical spacing (designed to separate sections which
may be empty) in printed file
- Update `FCodePrinter`
    - Extract interface functions from the Interface for Module printing
- Add scope handling to `FCodePrinter._constantImports` to ensure name
collision handling is correctly scoped.
- Update `PyCodePrinter`
- Extract interface functions from the Interface for the header printing
- Remove excessive vertical spacing (designed to separate sections which
may be empty) in printed file
- Call `_wrap(func)` in `_wrap_Interface` to wrap interface functions
(which are no longer wrapped by `_wrap_Module`
- Replace function insertion via `scope.functions[name] = func` with
`scope.insert_function(func, name)`
- Extract interface functions from the Interface for the declaration of
bind c functions in the C wrapper.
- Update `Scope`
    - Ensure docstrings specify if Python or low-level names are used
    - Return read-only dictionaries to prevent key misuse
- Add a new method `inline_variable_definition` to insert the definition
of a pseudo-variable used in an inline method
- Add assertions to ensure name is in the keys of `_used_symbols` which
maps Python names to low-level names
    - Add insert and remove methods instead of modifying dicts directly
        - `inline_variable_definition`
        - `insert_cls_construct`
        - `insert_function`
        - `remove_function`
    - Remove unused method `update_class`
- Update `SemanticParser`
- Ensure Python names are used when searching for objects in the scope
- Add `@overload` methods to scope via an Interface as they share a name
- Ensure functions are removed from the scope before they are replaced
with an updated version (e.g. prototype replaced with definition,
syntactic version replaced with semantic version, extra function added
to Interface)
    - Correct the scope of some variables/symbols
- Update `SyntacticParser`
- Ensure functions and classes are visited last (to ensure the parent
scope contains all symbol names)
- Ensure function and class names are collected early to complete module
scope symbols
- Add a `ReadOnlyDict` class to enforce correct usage of `Scope`
- Update example stub file as interfaces are now grouped together
- Move `GENERATED_NAME_COLLISION.py` test out of errors as it is no
longer failing

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>