Fix saving a `FunctionalFor` into a slice (#2428)
Saving a `FunctionalFor` into a slice did not work as it tried to
collect a non-existent allocation node and build an object (e.g. using
`append`). When the left-hand side is a slice no allocation is necessary
and the object should always be built using indexing. This PR makes 2
minor changes to ensure that this is correctly handled. Fixes #2096. It
also updates `sympy_to_pyccel` and `pyccel_to_sympy` so `//` can be used
in the size of the `FunctionalFor` object.