Install external libraries (#2401)
Handle STC as a fully external pre-compiled dependency. The dependency
is compiled on the fly to avoid needing to ship multiple libraries for
different compilers/OS systems. Fixes #2400. `Installer` classes are
created to install libraries cleanly. Each `Installer` has a
`install_to` method which returns a `CompileObj`. Previously the
`CompileObj` was a global object but this can lead to a race condition
when running 2 Pyccel instances which install the same library to
different folders.

**Commit Summary**
- Add a `compiler_family` property to `Compiler`
- Add a `is_debug` property to `Compiler`
- Make `Compiler._get_exec` into a public method `Compiler.get_exec` for
which the target language should be specified.
- Create `Installer` classes to install libraries cleanly
- `StdLibInstaller` to install internal libraries found in
`https://github.com/pyccel/pyccel/tree/devel/pyccel/stdlib`
    - `CWrapperInstaller`
    - `ExternalLibInstaller`
    - `STCInstaller`
    - `gFTLInstaller`
- Use `Installer` instances in `manage_dependencies`
- Ensure STC compilation files are installed