Fix shape/stride calculation (#2445)
Fix incorrect handling of shapes and strides of Fortran-order
multi-dimensional array that is C contiguous. Fixes #2440
If the array is a slice then the size must be calculated. The method of
doing this is different for C-ordered or F-ordered arrays. On `devel` we
use the array's flags to determine the ordering but if the shape is `(1,
N)` then the array is C-contiguous (can be accessed contiguously using
C-style indexing) even if it is F-ordered. This is therefore not a
reliable method of determining which size calculation method should be
used.