Support more argument types (#2389)
Currently Pyccel only handles positional arguments and keyword
arguments. It does not handle variadic arguments such as `*args` or
`**kwargs`. It also doesn't handle positional-only arguments or
keyword-only arguments. These concepts may be useful when wrapping
functions (see #2394). Further the lack of even basic support for these
concepts means that we cannot even parse our own code found in
https://github.com/pyccel/pyccel/blob/devel/pyccel/decorators.py which
may be used in translated functions. This issue is blocking for #2382 .

This PR adds parsing support for all these concepts and adds wrapping
support for positional-only arguments and keyword-only arguments.
Support is added up to the printing stage for `*args` and `**kwargs`. In
the case of `**kwargs` this is limited by #2057 #2058 (only the Python
translation is available).
`*args` cannot be supported out of the box with
`PyArg_ParseTupleAndKeywords`. Instead a custom loop would have to be
used to add this functionality manually.

**Commit Summary**
- Allow the creation of an empty tuple with a known class type.
- Add new arguments to `FunctionDefArgument` (`posonly`, `is_vararg`,
`is_kwarg`)
- Correct in/inout value of tuple argument
- Add `$` key to `PyArg_ParseTupleAndKeywords` string to split
positional only arguments from other arguments.
- Allow C to print a temporary tuple (to pass as a `*args` argument)
- Don't print `*args` as a keyword in Fortran
- Add handling of different argument types in Python
- Add default unused return values in `create_shared_library` to avoid a
backtrace if an error is raised in the wrapper
- Preserve `StarredArgs` in semantic stage so `*a` can pass `a` to
`*args` without unpacking `a` into a new temporary tuple
- Update logic for arguments to handle positional only and keyword only
arguments
- Add tests to ensure that the wrapper preserves the positional only and
keyword only properties in the arguments
- Add a test without a wrapper to ensure that the function call works
with `*args`
- Add `main` to C name clash handling