Fix excessive reallocation detection for lists, sets, and dicts (#2455)
Following #2454 the following code is falsely detected as attempting to
reallocate an array which is being used by another variable:
```python
def nested_dict_reassign():
    a = {1:[1,2,3], 2:[4,5,6]}
    print(a[1][0])
    a = {3:[1,2], 4:[4,5], 6:[7,8]}
    print(a[3][0])
```
This occurs because the print generates a temporary pointer. Such
temporaries should not be taken into account when finding problematic
pointers. This PR checks if the pointer associated with the target is a
temporary in order to discard these cases.