Fix extra tests due to bad autouse (#2468)
The new fixture `skipif_by_language` depends on `language` and uses
`autouse=True`. This means that the fixture is automatically used by
every test. These tests then depend on `language` even if this was not
the case previously. This results in the same test running 3 times
unnecessarily. The fix is to extract the language from the `request`
parameter rather than adding it as a dependency.