Deduce the name of methods in the semantic stage (#2410)
In the devel branch the names of methods are sometimes modified in the
codegen stage. This is not ideal for reasons discussed in #2405 . This
PR leverages the name collision mechanisms in the `Scope` class to
ensure that appropriate names are chosen earlier. The appropriate
renaming usually involves a prefix determined by the surrounding scope
and context. This PR adds a `scope_type` property to scopes to deduce
whether such prefixes are needed. The logic for the choice of prefix is
placed in the `NameClashChecker` subclasses. Fixes #2405

Renaming objects based on context means that it is trivial to rename
functions inside a module. This is done for C to introduce a namespace
per file. Fixes #1150 .

This PR makes renaming massively more common across the code base. As a
result it has exposed some remaining bugs where differences between
Python and low-level names were not taken into consideration.

**Commit Summary**
- Use only Python names for keys in `Module._internal_dictionary`
- Stop renaming imported functions
- Add `module_def_name` string parameter to `PyModule.__init__`,
accessible through a namesake property to ensure the name is used
consistently
- Stop renaming methods in C code printing
- Add `Scope.scope_type` property with value in `('module', 'function',
'class', 'loop', 'program')`. A similar `scope_type` argument is
required by `__init__` and `new_child_scope`
- Make `Scope.name` compulsory for all scope types except loops
- Modify `LanguageNameClashChecker.get_collisionless_name` function:
    - Add new arguments (prefix, context, parent_context)
- Use new arguments to decide if the prefix should be prepended to the
name
- Move `>> Calculating semantic annotations` print to more accurate
location
- Add `object_type` argument to Scope methods for name handling
    - Added to: `get_new_name`, and `insert_symbol`
    - `object_type` can be one of:
        - 'module'
        - 'function'
        - 'class'
        - 'variable' (default value)
- 'wrapper' : used to prevent renaming in the wrapper where names may be
compulsory for the API to work as expected
- Add `get_import_alias` method to `Scope` to handle expressions like
`import X as Y` without renaming the underlying object `X`
- Use `@low_level` decorator in `openmp.pyi` to avoid methods being
renamed with module prefix
- Update generated stub files in test
- Remove skip marks on C collision tests
- Reorder values in `set_methods.__all__` to fix pyccel_lint warning