Fix OpenMP handling in LLVM tests (#2473)
OpenMP tests are failing in the devel branch, however this is not a
problem with Pyccel, it is simply a problem with the installation of
LLVM on the runner. This PR fixes this which allows us to reactivate the
OpenMP tests.

Conversion tests have also been failing since we started using LLVM v21
(in v20 these were passing). This is related to overflow handling which
is undefined behaviour. These tests are therefore deactivated.

LLVM is also added to the PR test keys so it will be run when running
`/bot run pr_tests`. This will help prevent LLVM errors sneaking in
later.