Incorporate type modifiers into typing system (#2415)
Some type information does not directly represent a type but simply how
an object of this type behaves in certain circumstances. This is notably
the case for `Final` and `Optional`. Currently this information is
stored in the `Variable` class however this is not really sufficient
(particularly for `Optional`) as printing of other objects may need this
information (e.g. `dict.get` should be printed differently if `None` can
be returned).

Incorporating this information into the datatype is therefore necessary
for `Optional` but this is non-trivial as the code often relies on
statements like `isinstance(dtype, NativeIntegerType)` which should
return true if dtype is `Optional[NativeIntegerType]`. Many other
changes are required to make `Optional` work as a general concept so in
order to break up the work `Final` is first incorporated into the typing
system to work as a prototype. `Optional` will then be implemented
following this pattern.

In order to ensure that `isinstance(final_native_integer_type_obj,
NativeIntegerType)` returns true, `FinalNativeIntegerType` must inherit
from `NativeIntegerType`. In order for `FinalListNativeIntegerType` to
follow a similar pattern and to know how to initialise itself it is
simpler for all datatypes to be singletons. Previously things like
`HomogeneousListType` were `ArgumentSingleton`s meaning that there was 1
instance for each possible argument type. In order to create a singleton
with the correct inheritance a new class method `get_new` is added to
build the necessary type on the fly. `@cache` is used to ensure we don't
have multiple types representing the same concept.