Fix windows CI (#2488)
The installation of gFTL on Windows usually involves an installation of
m4 via these lines:

https://github.com/Goddard-Fortran-Ecosystem/gFTL/blob/5367705635d03d5722a9a4015892890ef25c6a07/CMakeLists.txt#L21-L22

however sourceforge has now moved these files. This causes the CI to
crash.
The repo for gFTL has since updated these lines to simply request that
users provide m4 in their installation:

https://github.com/Goddard-Fortran-Ecosystem/gFTL/blob/cb0415010ba7ce810ce0af33e4005ebd85e8c50c/CMakeLists.txt#L16-L17

This PR therefore follows that advice and installs m4.

The Windows installation process is also updated to use the new msys2
Software Distribution and Building Platform for Windows. This makes it
much simpler to install MPI. As a result the MPI wrappers are correctly
installed so it is no longer necessary to use `gfortran` with manually
set flags.