Fix passing non-contiguous slice (#2477)
Fix passing non-contiguous slice by improving the methods
`_check_pyarray_order` and `get_strides_and_shape_from_numpy_array`. It
is sufficient to check that the strides are correctly ordered to
determine the ordering. The shapes and strides then require 1 additional
piece of information in order to correctly pass the input to low-level
code. Fixes #2443.