Add support for zgbtrs (#2499)
Add support for Lapack's `zgbtrf` and `zgbtrs` in
`pyccel.stdlib.internals`. Add support for passing 1D arguments as `b`
argument to Lapack's `sgbtrs`, `dgbtrs`, and `zgbtrs` in
`pyccel.stdlib.internals`. Fixes #2497

Minimal testing exists for Lapack, see #925 and #1022. Fixing this is
left for a later PR. This PR only adds 1 line to a test to check that
Pyccel accepts the syntax