Handle allows_negative_indexes variables in list comprehension iterable (#2437)
Negative indices in a list comprehension expression (e.g. `[i for i in
x[-p-1:]]`) don't work because the shape of the indexed element is saved
incorrectly. This PR fixes that and ensures that an error is raised if
the size is too difficult to calculate (e.g. `y[a:b:c]` where a,b, and c
can each be either positive or negative). Fixes #2436

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>