Fix import class as renamed_class (#2498)
Fix expressions of the type `import A as B` where `A` is a class.
Previously the class definition was saved under the new name, but the
datatype wasn't leading to errors. This also meant that there was no way
to identify the appropriate imported name while printing. The datatype
(`cls_construct`) is now inserted into the scope under the correct name
and that name is retrieved during printing. Fixes #2496.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>