Fix duplicate imports in Python and stub files (#2505)
In the Python printer the `_additional_imports` dictionary is not needed
to store linker information and is the source of duplicate imports as
duplicates are avoided in this dictionary, but not between this
dictionary and the module imports. As the dictionary is unneeded a very
simple solution is to save the module imports in the
`_additional_imports` dictionary so that as new imports are added,
existing imports are found. Fixes #2291.