Stop using __new__ to simplify PyccelOperator types (#2493)
Stop using `__new__` to simplify PyccelOperator types. Instead use a
class method `make_simplified`. This avoids problems if the simplified
version has the same type as the original version (As can be the case
for expressions like `a+b+0`). This improves the readability of the
code, especially for array and slice sizes. Fixes #2486

This PR also adds several additional simplifications. Not all
simplifications are used but having a general implementation means we
don't have to keep revisiting this file when we have new use cases.