Fix optional `TypeVar` arguments (#2514)
Fix optional `TypeVar` arguments by leaving the type check for the
wrapper of the implementation of the function. Previously the wrapper of
the interface checked the type of all `TypeVar` arguments to determine
which implementation to call. Now it checks the type of all non-optional
`TypeVar` arguments. When the implementation has been chosen, all
arguments whose type has not yet been checked are checked for
correctness. Fixes #2484.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>