Fix bug when returning an expression that is translated to multiple lines (#2411)
Fix bug when returning an expression that is translated to multiple
lines. Fixes #2406

The code tries to avoid printing unnecessary assignments (e.g `return 1`
should be `return 1;` not `result = 1; return result;`). However the
case of `a & b` was not taken into account. This translates to multiple
expressions (Allocate, Assign, SetIntersectionUpdate). As the assignment
is not the last element it shouldn't be removed from the print.