Fix memory leak when memory is allocated in a function (#2461)
It is difficult to locate function memory allocation a posteriori.
However it is easy to identify while the call is being examined. In
order to fix #2292 we therefore add a new `alloc_type` to `Allocate`.
The `alloc_type` `'function'` indicates that the allocation should not
be printed, but any necessary deallocations should be. This `Allocate`
node is easy to locate later which allows us to rely on the old logic to
add missing deallocations. The argument `heap_mem_in_multirets` is also
added to `_ensure_inferred_type_matches_existing` to identify a variable
requiring heap allocation, which is collected as one of the returns in a
function returning multiple objects. This allows the functions to be
recognised without heap variables being falsely identified as pointers
(as previously they were seen to point to the function result element).
Valgrind tests are added to ensure proper deallocation.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>