Fix missing rpath for meson installation (#2523)
Ensure that libraries are installed to `__pyccel__/install` (for now
just the math library) so all internal libraries are found in the same
location. Add that location to the `install_rpath` of extension modules
and executables. This ensures that we correctly link to the math
library. This fixes errors of the kind:
```
ImportError: libpyc_math_f90.so: cannot open shared object file: No such file or directory
```

A floor division is added to a `pyccel-make` test to verify that the
installation works correctly.

Use static libraries to ensure this also works for Windows which has no
equivalent to `rpath`.