Aggregate Pyccel commands (#2522)
Aggregate the various Pyccel commands into one `pyccel` command with
multiple subcommands. The old commands are deprecated and should be
removed in the next Pyccel version. Fixes #2395

Some cleaning up of code is also done as a result. In particular 2 new
stages are added to the possible Pyccel stages. Catching errors is also
reorganised to avoid code duplication. The message describing the stage
at which the error occurred is now printed by the `Errors` singleton
itself. This allows all errors and warnings to be displayed together in
one command after executing the workflow.

**Commit Summary**
- Fix installation of MPI for flang
- Add two new Pyccel stages after `syntactic`, `semantic`, `codegen`,
and `wrapper`:
    - `compilation`
    - `buildgen`
- Remove try/except blocks from pipelines. Place them round the pipeline
calls to avoid code duplication
- Call `print(errors)` and `pyccel_stage.pyccel_finished()` after 
- Simplify commands so as much as possible is handled by argparse.
- Add new file `pyccel/commands/argparse_helpers.py` to hold helper
functions for this purpose
    - Move some trivial commands into related pipelines
- Use `**vars(parser.parse_args())` to pass `kwargs`. This ensures no
kwargs are ignored
- Add a new `console.py` file which just dispatches the `pyccel` command
to the relevant subcommands.
- Rename old `console.py` file to `pyccel_compile.py` to describe the
`pyccel compile` sub-command.
- Add a `pyccel config` command as suggested in #2370 (the check option
is not added so the issue is not closed)
- For each sub-command add a `setup_pyccel_CMD_parser` function to add
necessary arguments to a parser
- For each sub-command add a `pyccel_CMD` function to import and call
the pipeline
- For each deprecated command use `setup_pyccel_CMD_parser` and
`pyccel_CMD` to preserve functionality. Add deprecation message
- Remove `Errors.check()` function which does nothing except add an
extra function call
- Add an error if no files are passed to `pyccel make`
- Print files being translated by `pyccel make` at all levels of
verbosity
- Remove private functions `handle_error(s)` from `execute_pyccel`,
`execute_pyccel_make`, and `execute_pyccel_wrap`
- Remove `output_name` from parameters of `execute_pyccel_wrap` (name of
generated module defaults to stem of input filename)
- Fix meson Fortran linking by specifying the linker language
- Fix meson symbol visibility (meson makes symbols hidden leading to
linker errors)

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>