Stop using deprecated commands in tests (#2526)
Following #2522 commands such as `pyccel-make` are deprecated in favour
of `pyccel make`. The calls to these CLI tools in the tests were not
removed to ensure coverage while working on #2522. Now that #2522 is
merged we don't want our tests to show deprecation warnings so the tools
used in the tests are updated.

In addition, we update some docstrings and rename the workflow
`linux_pyccel-test_cmd` to `linux_pyccel_test`.

There is also a minor bug fix to the new `pyccel config` command so it
doesn't crash if the new json file doesn't exist yet.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>